Added gradient_box_splitter
The splitter calculates sums of gradient absolute values in each
direction (X,Y,Z) and selects the dimension with largest gradient value
to split along. Then repeats the process for the selected axis until it
reaches half the value of the total summed gradient for this axis.

It
should mean, that there is approximately similar amount of complexity on
both sides of the split.